Send SIGHUP when stream ends
Previously, we were sending the SIGHUP when the socket was disconnected.
The socket was kept alive and reused for all tutorials.  Only once the
top-level react component was unmounted, was the socket disconnected.

Now, we make use of each tutorial having its own stream and HUP the tty
when the stream ends.  This results in the sessions being cleaned up
much sooner.